// zklink-web-simple/src/styles/constants/type.ts
export interface ColorType {
  bgGray: string // --bg-color-light-gray, --bg-color-dark-grey
  bg: string // --bg-color-light-white, --bg-color-dark-black
  bgLightGray: string // --bg-color-light-light-grey, --bg-color-dark-dark-gray
  bgPopup: string // --bg-color-light-popup-bg, --bg-color-dark-popup-bg
  primary10: string // --primary-color-light-10, --primary-color-dark-10
  primary20: string // --primary-color-light-20, --primary-color-dark-20
  primary30: string // --primary-color-light-30, --primary-color-dark-30
  primary40: string // --primary-color-light-40, --primary-color-dark-40
  primary50: string // --primary-color-light-50, --primary-color-dark-50
  primaryGreyGreen: string // --primary-color-light-grey-green, --primary-color-dark-grey-green
  textHighlight: string // --text-color-light-highlight, --text-color-dark-highlight
  text00: string // --text-color-light-00, --text-color-dark-00
  text10: string // --text-color-light-10, --text-color-dark-10
  text20: string // --text-color-light-20, --text-color-dark-20
  text30: string // --text-color-light-30, --text-color-dark-30
  text40: string // --text-color-light-40, --text-color-dark-40
  text50: string // --text-color-light-50, --text-color-dark-50
  text60: string // --text-color-light-60, --text-color-dark-60
  text70: string // --text-color-light-70, --text-color-dark-70
  text80: string // --text-color-light-80, --text-color-dark-80
  text90: string // --text-color-light-90, --text-color-dark-90
  text100: string // --text-color-light-100, --text-color-dark-100
  buy10: string // --buy-light-10, --buy-dark-10
  buy20: string // --buy-light-20, --buy-dark-20
  buy30: string // --buy-light-30, --buy-dark-30
  buy40: string // --buy-light-40, --buy-dark-40
  sell10: string // --sell-light-10, --sell-dark-10
  sell20: string // --sell-light-20, --sell-dark-20
  sell30: string // --sell-light-30, --sell-dark-30
  sell40: string // --sell-light-40, --sell-dark-40
  notificationBlue01: string // --notifications-light-blue01, --notifications-dark-blue01
  notificationBlue02: string // --notifications-light-blue02, --notifications-dark-blue02
  notificationGreen01: string // --notifications-light-green01, --notifications-dark-green01
  notificationGreen02: string // --notifications-light-green02, --notifications-dark-green02
  notificationRed01: string // --notifications-light-red01, --notifications-dark-red01
  notificationRed02: string // --notifications-light-red02, --notifications-dark-red02
  notificationYellow01: string // --notifications-light-yellow01, --notifications-dark-yellow01
  notificationYellow02: string // --notifications-light-yellow02, --notifications-dark-yellow02
  notificationPurple01: string // --notifications-light-purple01, --notifications-dark-purple01
  notificationPurple02: string // --notifications-light-purple02, --notifications-dark-purple02

  bgMask: string
  bgHistoryTabs: string
  bgNavButton: string
  networkSelectionBgColor: string
  withdrawTypeBgColor: string
  TabHoverColor: string
  balanceGroupBorderColor: string
  balanceGroupHoverBgColor: string
  balanceGroupHoverBorderColor: string

  DatePickerInput: string

  BoxBg01: string
  BoxBg02: string
  BoxBg03: string
  BoxBg04: string
  BoxBorder01: string
  BoxLineBg: string
  BoxBtnBg: string
  BoxBtnBg02: string
  BoxBtnColor: string
  BoxRankBg01: string
  BoxRankBorder01: string
  BoxRankBg02: string
  BoxRankBorder02: string
  BoxRankBg03: string
  BoxRankBorder03: string
  BoxDialogBg: string
  BoxDialogBg02: string
  BoxHistoryRow: string
}